+++ src/main/java/com/programmers/jpa/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.programmers.jpa.domain;
+package com.programmers.jpa.repository;
 
+import com.programmers.jpa.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
