+++ src/test/java/com/programmers/jpa/domain/MemberRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.programmers.jpa.domain;
 
+import com.programmers.jpa.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
